+++ doc/release-notes.md
@@ -7,3 +7,4 @@ Bitcoin ABC version 0.26.6 is now available from:
 This release includes the following features and fixes:
  - The `getblock` RPC verbose mode level 2 now returns the transactions fees if
    undo data is available. See the RPC documentation for more details.
+ - Add a checkpoint after the Nov. 15th, 2022 eCash upgrade.